requirements.txt
django
pillow
django-storages[azure]